+++ src/tilda/types/Type_CharCollectionNull.java
@@ -19,7 +19,7 @@ package tilda.types;
 
 public class Type_CharCollectionNull extends Type_CharCollection implements Nullable
   {
-    public Type_CharCollectionNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_CharCollectionNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
